+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/nodeviews/mediaNodeUpdater.spec.ts
@@ -14,7 +14,6 @@ import {
   MediaNodeUpdater,
   MediaNodeUpdaterProps,
 } from '../../../../../plugins/media/nodeviews/mediaNodeUpdater';
-import * as mediaCommon from '../../../../../plugins/media/utils/media-common';
 import { MediaProvider } from '../../../../../plugins/media/pm-plugins/main';
 
 describe('MediaNodeUpdater', () => {
@@ -30,9 +29,6 @@ describe('MediaNodeUpdater', () => {
       .mockReturnValue(() => true);
     jest.spyOn(commands, 'updateMediaNodeAttrs').mockReturnValue(() => true);
     jest.spyOn(commands, 'replaceExternalMedia').mockReturnValue(() => true);
-    jest
-      .spyOn(mediaCommon, 'getViewMediaClientConfigFromMediaProvider')
-      .mockReturnValue(Promise.resolve(getDefaultMediaClientConfig()));
 
     const contextIdentifierProvider: Promise<
       ContextIdentifierProvider