+++ packages/editor/editor-core/src/labs/next/presets/types.ts
@@ -1,4 +1,10 @@
+import { EditorPlugin } from '../../../types/editor-plugin';
+
 export interface EditorPresetProps {
   excludes?: Array<string>;
   experimental?: Array<string>;
 }
+
+export type PluginsPreset = Array<
+  ((arg: any) => EditorPlugin) | [(arg: any) => EditorPlugin, object]
+>;