+++ packages/editor/editor-core/src/plugins/fake-text-cursor/styles.ts
@@ -1,6 +1,4 @@
-// @ts-ignore: unused variable
-// prettier-ignore
-import { css, Styles, StyledComponentClass } from 'styled-components';
+import { css } from 'styled-components';
 import { colors } from '@atlaskit/theme';
 
 export const fakeCursorStyles = css`
@@ -9,7 +7,6 @@ export const fakeCursorStyles = css`
       display: inline;
       pointer-events: none;
       position: relative;
-      height: 15px;
     }
 
     .ProseMirror-fake-text-cursor::after {
@@ -17,7 +14,6 @@ export const fakeCursorStyles = css`
       display: inline;
       top: 0;
       position: absolute;
-      height: 100%;
       border-right: 1px solid rgba(0, 0, 0, 0.4);
     }