+++ packages/media/media-test-helpers/src/mockData/index.ts
@@ -1,6 +1,6 @@
 import { MockURL } from 'xhr-mock/lib/MockURL';
 import { MockHeaders } from 'xhr-mock/lib/types';
-import { ClientBasedAuth } from '@atlaskit/media-store';
+import { ClientBasedAuth } from '@atlaskit/media-core';
 import { RECENTS_COLLECTION } from '@atlaskit/media-client/constants';
 
 export * from './matchers';