+++ packages/media/media-viewer/src/newgen/analytics/index.ts
@@ -1,5 +1,4 @@
-import { FileState } from '@atlaskit/media-client';
-import { MediaType } from '@atlaskit/media-store';
+import { FileState, MediaType } from '@atlaskit/media-client';
 import {
   name as packageName,
   version as packageVersion,